+++ Westwind.AspNetCore.Markdown/MarkdownParser/IMarkdownParserFactory.cs
@@ -1,0 +1,7 @@
+﻿namespace Westwind.AspNetCore.Markdown
+{
+    public interface IMarkdownParserFactory
+    {        
+        IMarkdownParser GetParser(bool usePragmaLines = false, bool forceLoad = false);
+    }
+}