+++ Westwind.AspNetCore.Markdown/MarkdownParser/MarkdownParserFactory.cs
@@ -3,14 +3,12 @@
     /// <summary>
     /// Retrieves an instance of a markdown parser
     /// </summary>
-    public static class MarkdownParserFactory
+    public class MarkdigMarkdownParserFactory : IMarkdownParserFactory
     {
-        public static string DefaultMarkdownParserName { get;  } = "MarkDig";
-
         /// <summary>
         /// Use a cached instance of the Markdown Parser to keep alive
         /// </summary>
-        public static IMarkdownParser CurrentParser;
+        private IMarkdownParser CurrentParser;
 
 
         /// <summary>
@@ -20,18 +18,16 @@
         /// <param name="usePragmaLines">If true adds pragma line ids into the document that the editor can sync to</param>
         /// <param name="parserAddinId">optional addin id that checks for a registered Markdown parser</param>
         /// <returns>Mardown Parser Interface</returns>
-        public static IMarkdownParser GetParser(bool usePragmaLines = false,
-                                                bool forceLoad = false)                                                
+        public IMarkdownParser GetParser(bool usePragmaLines = false,
+                                         bool forceLoad = false)
         {
             if (!forceLoad && CurrentParser != null)
                 return CurrentParser;
-            
+
             CurrentParser = new MarkdownParserMarkdig(usePragmaLines, forceLoad);
             return CurrentParser;
         }
-        
-        
     }
 
-    
+
 }